+++ src/primitives.rs
@@ -1,0 +1,6 @@
+use crate::Sphere;
+
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd)]
+pub enum Primitive {
+    Sphere(Sphere)
+}